Permisiuni pe date, facute programatic la botul calului

+++ spring-one/src/main/java/victor/training/spring/web/controller/TrainingController.java
@@ -3,12 +3,17 @@ package victor.training.spring.web.controller;
 import java.text.ParseException;
 import java.util.EnumMap;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import victor.training.spring.web.controller.dto.TrainingDto;
+import victor.training.spring.web.domain.Training;
+import victor.training.spring.web.repo.TrainingRepo;
+import victor.training.spring.web.security.SecurityUser;
 import victor.training.spring.web.service.TrainingService;
 
 import static java.util.Arrays.asList;
@@ -46,11 +51,22 @@ public class TrainingController {
 	// TODO [SEC] 1 Allow only for ROLE 'ADMIN'
 	// TODO [SEC] 2 Authorize the user to have the authority 'deleteTraining'
 	// TODO and @accessController.canDeleteTraining(#id)
+
+	// un user are voie sa stearga un curs <=> training.teacher.id IN user.managedTheacherIds
 	/** @see victor.training.spring.web.domain.UserProfile */
 	@DeleteMapping("{id}")
 	public void deleteTrainingById(@PathVariable Long id) {
+		SecurityUser user = (SecurityUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		Training training = trainingRepo.findById(id).get();
+		if (!user.getManagedTeacherIds().contains(training.getTeacher().getId())) {
+			throw new IllegalArgumentException("n-ai voie!");
+		}
+
 		trainingService.deleteById(id);
 	}
+
+	@Autowired
+	private TrainingRepo trainingRepo;
 
 	@PostMapping
 	public void createTraining(@RequestBody TrainingDto dto) throws ParseException {
